【BUG修复】 修复了只能接受到一种消息的问题

+++ 01. GCB测试工具/04. 工程程序/01. 工程程序/GCBTestTool/GCBTestTool/ChildDialog/GCBMainDlg.cpp
@@ -208,7 +208,7 @@ void GCBMainDlg::StartDraw(int ControlID)
 	int oldMinX = 0, oldMinY = pitureHeight / 2;
 
 	list<float>::iterator iter = this->showValueLst.begin();
-	for (int i = 1; i < min(PICTRUE_SHOW_SIZE, this->showValueLst.size()); ++i, iter++) {
+	for (int i = 1; i < (int) min(PICTRUE_SHOW_SIZE, this->showValueLst.size()); ++i, iter++) {
 		double showMax = 0, showMin = 0;
 
 		showMax = *iter;
